Use ca certificate from DCOS_SSL_VERIFY.
Summary:
Shakedown reads the path to the ca certificate for SSL connections from `DCOS_SSL_VERIFY`. This
changes enables SSL with the given certificate for all connections.

JIRA issues: MARATHON-8423
